+++ kaizo/utils/__init__.py
@@ -1,7 +1,13 @@
+from .common import split_by_first_dot
+from .entry import DictEntry, Entry, FieldEntry, ListEntry, ModuleEntry
 from .fn import FnWithKwargs
-from .parser import ConfigParser
 
 __all__ = (
-    "ConfigParser",
+    "DictEntry",
+    "Entry",
+    "FieldEntry",
     "FnWithKwargs",
+    "ListEntry",
+    "ModuleEntry",
+    "split_by_first_dot",
 )